feat: issue mega clipboard
Copies list of all issues for current sprint onto the clipboard.
